+++ quiltplus/package.py
@@ -4,7 +4,7 @@ import logging
 import shutil
 
 from quilt3 import Package  # type: ignore
-from typing_extensions import Self, Type
+from typing_extensions import Self, Type, Coroutine
 
 from .local import QuiltLocal
 from .uri import QuiltUri
@@ -47,9 +47,13 @@ class QuiltPackage(QuiltLocal):
 
     async def list(self, opts: dict = {}):
         return [self.path_uri(k) for k in await self.child()]
+    
+    def stage_uri(self, stage: str, sub_path: str):
+        return self.path_uri(sub_path).replace("quilt+",f"quilt+stage+{stage}")
 
     async def diff(self, opts: dict = {}):
-        pass
+        _diff = self._diff()
+        return [self.stage_uri(k, v) for k, sublist in _diff.items() for v in sublist]
 
     async def get(self, opts: dict = {}):
         dest = self.check_path(opts)